Fixed minor bugs

+++ Lab05/SpriteComponent.cpp
@@ -29,8 +29,8 @@ void SpriteComponent::Draw(SDL_Renderer* renderer)
 		r.y = static_cast<int>(mOwner->GetPosition().y - r.h / 2);
         
         //Adjust position according to cam position
-        r.x -= mOwner->GetGame()->GetCamPosition().x;
-        r.y -= mOwner->GetGame()->GetCamPosition().y;
+        r.x -= static_cast<int>( mOwner->GetGame()->GetCamPosition().x);
+        r.y -= static_cast<int>( mOwner->GetGame()->GetCamPosition().y);
 
 		// Draw (have to convert angle from radians to degrees, and clockwise to counter)
 		SDL_RenderCopyEx(renderer,
